+++ src/components/account/edit/sections/UserSecuritySection.vue
@@ -61,8 +61,8 @@
 </template>
 
 <script lang="ts">
-    import { useModelWrapper } from "@/use/ModelWrapper";
-    import ErrorBag from "@/use/ErrorBag";
+    import { useModelWrapper } from "@/use/helpers/ModelWrapper";
+    import ErrorBag from "@/use/helpers/ErrorBag";
 
     export default {
         name: "RoleSection",
