+++ src/components/course/edit/sections/TimeSection.vue
@@ -39,8 +39,8 @@
 </template>
 
 <script lang="ts">
-    import ErrorBag from "@/use/ErrorBag";
-    import { useModelWrapper } from "@/use/ModelWrapper";
+    import ErrorBag from "@/use/helpers/ErrorBag";
+    import { useModelWrapper } from "@/use/helpers/ModelWrapper";
 
     export default {
         name: "RestrictionsSection",
