+++ src/components/course/list/common/CourseList.vue
@@ -27,7 +27,7 @@
     import LecturerCourse from "../lecturer/LecturerCourse.vue";
     import StudentCourse from "../student/StudentCourse.vue";
     import CourseManagement from "@/api/CourseManagement";
-    import GenericResponseHandler from "@/use/GenericResponseHandler";
+    import GenericResponseHandler from "@/use/helpers/GenericResponseHandler";
     import Course from "@/api/api_models/course_management/Course";
     import APIResponse from "@/api/helpers/models/APIResponse";
     import { computed, ref, onBeforeMount, watch } from "vue";
