+++ src/components/course/list/lecturer/LecturerCourse.vue
@@ -42,7 +42,7 @@
     import router from "@/use/router";
     import Course from "@/api/api_models/course_management/Course";
     import UserManagement from "@/api/UserManagement";
-    import ProfileResponseHandler from "@/use/ProfileResponseHandler";
+    import ProfileResponseHandler from "@/use/helpers/ProfileResponseHandler";
     import Lecturer from "@/api/api_models/user_management/Lecturer";
 
     export default {
