+++ src/components/modals/EnterPasswordModal.vue
@@ -28,7 +28,7 @@
     import Modal from "@/components/modals/Modal.vue";
     import UserManagement from "@/api/UserManagement";
     import { ref } from "vue";
-    import LoginResponseHandler from "@/use/LoginResponseHandler";
+    import LoginResponseHandler from "@/use/helpers/LoginResponseHandler";
     import { useStore } from "@/use/store/store";
 
     export default {
