+++ src/components/settings/SecuritySection.vue
@@ -94,7 +94,7 @@
     import { store } from "@/use/store/store";
     import EnterPasswordModal from "@/components/modals/EnterPasswordModal.vue";
     import UserManagement from "@/api/UserManagement";
-    import GenericResponseHandler from "@/use/GenericResponseHandler";
+    import GenericResponseHandler from "@/use/helpers/GenericResponseHandler";
     import AuthenticationManagement from "../../api/AuthenticationManagement";
 
     export default {
