+++ src/use/helpers/LoginResponseHandler.ts
@@ -1,5 +1,5 @@
 import ResponseHandler from "./ResponseHandler";
-import APIResponse from "../api/helpers/models/APIResponse";
+import APIResponse from "@/api/helpers/models/APIResponse";
 
 export default class LoginResponseHandler implements ResponseHandler<boolean> {
     handleReponse(response: APIResponse<boolean>): boolean {
