+++ src/use/helpers/ProfileResponseHandler.ts
@@ -1,5 +1,5 @@
 import ResponseHandler from "./ResponseHandler";
-import APIResponse from "../api/helpers/models/APIResponse";
+import APIResponse from "@/api/helpers/models/APIResponse";
 import Student from "@/api/api_models/user_management/Student";
 import Lecturer from "@/api/api_models/user_management/Lecturer";
 import Admin from "@/api/api_models/user_management/Admin";
