+++ src/views/common/Login.vue
@@ -69,10 +69,10 @@
 <script lang="ts">
     import Router from "@/use/router/";
     import { useStore, store } from "@/use/store/store";
-    import { Role } from "../../entities/Role";
+    import { Role } from "@/entities/Role";
     import UserManagement from "@/api/UserManagement";
     import { ref, onMounted } from "vue";
-    import LoginResponseHandler from "@/use/LoginResponseHandler";
+    import LoginResponseHandler from "@/use/helpers/LoginResponseHandler";
 
     export default {
         components: {},
