+++ src/views/common/Welcome.vue
@@ -27,7 +27,7 @@
     import { useStore } from "@/use/store/store";
     import { ref, onBeforeMount } from "vue";
     import LoadingSpinner from "@/components/common/loading/Spinner.vue";
-    import { checkPrivilege } from "@/use/PermissionHelper";
+    import { checkPrivilege } from "@/use/helpers/PermissionHelper";
     import { Role } from "@/entities/Role";
 
     export default {
