+++ src/views/lecturer/LecturerCourseList.vue
@@ -35,7 +35,7 @@
     import { ref } from "vue";
     import CourseTypeFilter from "@/components/course/list/common/CourseTypeFilter.vue";
     import { CourseType } from "@/entities/CourseType";
-    import { checkPrivilege } from "@/use/PermissionHelper";
+    import { checkPrivilege } from "@/use/helpers/PermissionHelper";
     import { Role } from "@/entities/Role";
 
     export default {
